All files moved to separate folders
