+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './styles.scss'
 import { createBrowserRouter, RouterProvider, Outlet } from 'react-router-dom';
 import Header from './components/Header';
-import Home from './pages/Home';
+import Home from './pages/Home/Home';
 import Cenovnik from './pages/Cenovnik';
 import Kontakt from './pages/Kontakt';
