+++ src/pages/Home/Styles/Faq.scss
