+++ src/pages/Home/components/InfoSection.tsx
